\name{bxplot.JGR}
\alias{bxplot.JGR}
\title{Dialog box for generating boxplots for data with factor subsets}
\description{
	Creates boxplots for chosen subsets of the data. A boxplot displays
	distributional charactersitics (location and shape) of the data. Boxplots
	also provide a qualitative means of comparing several data sets.
}

\details{
	
	\bold{User Interface}

	Select \emph{Graph -> Boxplot} from the menus.  A dialog box will open.
	
	Select the data set of interest from the pull-down menu, or
	browse for a tab-delimited text file.
	
Select the variable that you wish to plot from the pull-down menu, \bold{Result}.
By default, the module will produce two plots, one in the original units
of Result, and one in log transformed units.

Text indicating sample size of the data set is included in the plot if \bold{Sample Sizes} is selected.
You can alter the labels for the plot by modifying \bold{Plot Title} and
\bold{Result-Axis}.

By default, the labels on the x-axis are printed horizontally, but can
be changed to vertical with \bold{Rotate X Axis Labels}.

}



\keyword{hplot} 

