\name{pca.fa.JGR}
\alias{pca.fa.JGR}
\title{Dialog box for calculating a principle component analysis}
\description{
  This function receives inputs from the Java GUI for factor
  analysis/principle component analysis
 and sends these parameters to the FA/PCA function.
}

\details{

  \strong{User Interface}

  Select \emph{Analysis Tools -> Factor Analysis} from the menus.  A dialog
  box will open.  Select the data set of interest from the pull-down
  menu, or browse for a tab-delimited text file.  The Data Subsetting
  tab can be used to select a subset of the data file by choosing a
  variable from the pull down menu and then selecting the levels of that
  variable to include.  You can hold down the <CTRL> key to select
  several levels.
  
Select the \bold{No. of Factors} you wish to compute.
Select all explanatory (\bold{Variables}) you wish to include in
the analysis.  You can hold down the <CTRL> key to select several
variables. 

Output options include generating a \bold{Scree Plot},
\bold{Loading/Score Plot}, or using \bold{Rotated Variables}.

Analysis options include centering all variables to zero and scaling all
variables to a unit variance.
	
}


\keyword{hplot} 

