\name{refids.or}
\alias{refids.or}
\docType{data}
\title{Reference site identifications western OR}
\description{
  Sample identifiers for reference site in western Oregon
}
\format{
  A data frame with 245 observations on the following 13 variables.
  \describe{
    \item{\code{SITE.ID}}{Unique sample code}
    \item{\code{REF}}{Flag for reference site}
  }
}
\source{
  Oregon Department of Environmental Quality
}
\keyword{datasets}
